backend/requirements.txt
asgiref == 3.7.2
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
transliterate
django-filter
drf-base64-filename
drf_yasg
djoser
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0
